correction de la connexion
